find error in functional tests
