+++ src/functionalTest/java/co/com/bancolombia/PluginCleanFunctionalTest.java
@@ -943,6 +943,31 @@ public class PluginCleanFunctionalTest {
     assertEquals(result.task(":" + task).getOutcome(), TaskOutcome.SUCCESS);
   }
 
+  @Test
+  public void shouldUpdateProject() {
+    canRunTaskGenerateStructureWithOutParameters();
+
+    String task = "updateCleanArchitecture";
+
+    runner.withArguments(task);
+    runner.withProjectDir(projectDir);
+    BuildResult result = runner.build();
+
+    assertEquals(result.task(":" + task).getOutcome(), TaskOutcome.SUCCESS);
+  }
+
+  @Test
+  public void shouldUpdateProjectWithOneDependency() {
+    canRunTaskGenerateStructureWithOutParameters();
+    String task = "updateCleanArchitecture";
+
+    runner.withArguments(task, "--dependencies=org.mockito:mockito-core org.projectlombok:lombok");
+    runner.withProjectDir(projectDir);
+    BuildResult result = runner.build();
+
+    assertEquals(result.task(":" + task).getOutcome(), TaskOutcome.SUCCESS);
+  }
+
   private void writeString(File file, String string) throws IOException {
     try (Writer writer = new FileWriter(file)) {
       writer.write(string);
